Make core a bit more reusable.
